requirements.txt
pyperclip
requests
ruamel.yaml
